transfer clifford synthesis functions from quantum_info to synthesis
